stacks/cells/genome/requirements.txt
# AIOS Genome Cell Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0
